:calendar: Calendar integration (#1191)
* :construction: Calendar integration

* :construction: Use hints instead of visibility

* :triangular_ruler: Use Y offset instead of height

* :warning: Remove warning

* :bug: Make calendar sober

* :runner: Must proceed at high velocity

* :scroll: Hack explanation

* :green_heart:

* :green_heart:

* :green_heart:

+++ Toggl.Daneel/ReportsCalendarCollectionViewLayout.cs
@@ -9,7 +9,7 @@ namespace Toggl.Daneel
     public class ReportsCalendarCollectionViewLayout : UICollectionViewLayout
     {
         public const float CellHeight = 42;
-        private const float columnCount = 7;
+        private const int columnCount = 7;
         private nfloat cellWidth;
         private UICollectionViewLayoutAttributes[] attributes;
